refactor: CarName 원자값 포장 처리
