+++ src/main/java/step5/domain/Car.java
@@ -3,9 +3,7 @@ package step5.domain;
 import step5.domain.result.CarMovementResult;
 
 public class Car {
-    private static final int CAR_NAME_MAX_LENGTH = 5;
-
-    private final String name;
+    private final CarName name;
     private int moveCount;
 
     public static Car withName(String name) {
@@ -13,14 +11,7 @@ public class Car {
     }
 
     private Car(String name) {
-        assertNameLengthUnderFive(name);
-        this.name = name;
-    }
-
-    private void assertNameLengthUnderFive(String name) {
-        if (CAR_NAME_MAX_LENGTH < name.length()) {
-            throw new IllegalArgumentException("자동차 이름은 5자를 초과할 수 없습니다.");
-        }
+        this.name = new CarName(name);
     }
 
     public CarMovementResult move(Moving movingStrategy) {
