+++ src/main/java/step5/domain/result/CarMovementResult.java
@@ -1,10 +1,16 @@
 package step5.domain.result;
 
+import step5.domain.CarName;
+
 public class CarMovementResult {
-    private final String carName;
+    private final CarName carName;
     private final int moveCount;
 
     public CarMovementResult(String carName, int moveCount) {
+        this(new CarName(carName), moveCount);
+    }
+
+    public CarMovementResult(CarName carName, int moveCount) {
         this.carName = carName;
         this.moveCount = moveCount;
     }
@@ -14,6 +20,6 @@ public class CarMovementResult {
     }
 
     public String carName() {
-        return carName;
+        return carName.value();
     }
 }
